requirements.txt
fastapi[all]==0.115.8
uvicorn==0.34.0
gunicorn==21.2.0
sqlalchemy~=2.0.37
asyncpg
alembic~=1.14.1
python-dotenv~=1.0.1
redis<5.0.0,>=4.2.0
fastapi-cache2[redis]==0.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0